Fixed the new bug
From previous commit there was a new bug dealing with a certain user not being able to edit the two starter homes. Found out that there was a copy of one user in the database. Took the copy out of the database and placed all the homes in the perspective order.